fix transition tests race condition DO NOT MERGE
Transition tests are reliant upon layout running, which sets various
values that are then checked The mechanism used to ensure that layout
had run (Instrumentation.waitForIdleSync()) doesn't do what we need, and
it is possible for the idle condition to succeed even while the queue still
has the pending layout request on it.

The fix is to wait on an actual layout to run instead.

Issue #31249295 CTS: ChangeBoundsTest failures

Change-Id: Iec7a5735a9d86fbe98c58acd1f32e325e9fdcab7