Fix domains in domain-config test
The domain-config test was overly pedantic about android.com only (so
that it could test not matching all subdomains). This broke with
redirects android.com redirecting to www.android.com which was outside
the domain-config.

This commit adds www.android.com to the domain-config, long term we
should look at getting a static target for these tests that is more
minimal but meets all the needs for these tests.

Bug: 33726200
Test: cts-tradefed run cts -m CtsNetSecConfigBasicDomainConfigTestCases
-t android.security.net.config.cts.DomainConfigTest#testDomainConfig

Change-Id: I12ec80fe1ce2ea9ef22966d70651f0119c12f322