Opt UI_MODE_TYPE_TELEVISION out of SearchManagerTest
Cherry-picked c95dc0877769201a7368bf2feb1cfbb0edb64a9f from internal
branch.

The Activity#onSearchRequested() API method is documented to not be
supported when the app is running in a UI_MODE_TYPE_TELEVISION. Hence,
topSearch(), setOnDismissListener() and setOnCancelListener() would
also not work in such scenarios.

The test is skipped when running on a device with UI_MODE_TYPE_TELEVISION.

BUG: 33040372
Change-Id: Ica5a1fa3b18eaa2d2564e71c454e2e2e5eb096d9