Add tests to verify activity ordering behavior.
This changelist adds two tests which ensures the correct behavior when
using the FLAG_ACTIVITY_REORDER_TO_FRONT flag. The first ensures that
the correct activity history is present when moving to and from a
root activity. The second exercises that the history and focus is
correct for an activity when moving to and from the home stack.

This changelist also adds a builder for launching an activity from
LaunchingActivity. Additionaly, an additional activity,
AltLaunchingActivity, has been added so that we can return to the
LaunchingActivity in the same manner.

Bug: 25784754
Bug: 25487874
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerActivityVisibilityTests#testReorderToFrontBackstack
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerActivityVisibilityTests#testReorderToFrontChangingStack
Change-Id: I482a687bd2bb8cd3800517bafc2ba47c17a7f625