Verify that IMAGE_CAPTURE writes to content://.
This test enforces the IMAGE_CAPTURE contract, which says the camera
should write data into EXTRA_OUTPUT when provided, especially when
it's a content:// Uri.

We can't rely on any particular camera app picking up the phone at
the other end, so we grope around to try taking a photo.  The CAMERA
keycode is documented to take a photo, and DPAD_CENTER is a
reasonable proxy for confirmation of the photo.  If those fail, we
fall back to looking for an explicit view ID.

Test: cts-tradefed run commandAndExit cts-dev -m CtsProviderTestCases -t android.provider.cts.MediaStoreUiTest#testImageCapture
Bug: 27981143, 16344345
Change-Id: Ie8f7017dde97ceb327196cbbc3d67a7f8c6335c5