Create @VendorInterfaceTest annotation and provide examples
The annotation can be applied at the testcase and the test class
level. The annotation is not sufficient for signalling that the
test should be run as part of cts-vendor-interface, and an
include-filter for the module (e.g. CtsAccelerationTestCases)
is also necessary.

Fixes: 36423839
Test: run cts-vendor-interface
Change-Id: I63c11c6f7f3a57ec2f1264b1c34d7a0e8bb93363
(cherry picked from commit 2a4efa91f237287bc0d9e859119344e895057f4d)