Retry starting autofill to deal with race
Telling the activity that an autofillservice was enabled is
unsyncronized with the rest of the test. Hence it can happen that this
notification comes after the focus is set. This means, that the focus
does not trigger an autofill anymore.

Test: cts-tradefed run cts-dev -m CtsAutoFillServiceTestCases
Change-Id: Id8a50b0efad1d4104245a0d1556a54ccd9820eba
Fixes: 36385395
(cherry picked from commit 19ee1f7cf1b52a4860d5d7032c12e3a871744032)