Update bombfont to reproduce the security issue after I24aa49860790
After I24aa49860790, Minikin starts using the first table if there
are multiple tables with the same platform and encoding IDs. In past
the last table was used. The table selection from the same encoding
table is an undefined behavior since spec says that the same encoding
table should appear only once.

This change breaks TypefaceTest.testInvalidCmapFont/2 test cases.
The bombfont had two same (3,1) encoding table. The first table had
no problem, but the second table had an intentional problem. The test
makes sure that the font is rejected due to Minikin being unable to
read the second table. However, once Minikin start reading the first
table by I24aa49860790, the problematic table won't be read and the
font won't be rejected.

The test was originally introduced for rejecting the font due to
invalid cmap format 12 table, so table selection is not a part of test
expectation. So, to read the second table, this CL changes the
encoding of the second table from (3,1) to (3,10). (3,10) encoding
has higher priority than (3,1), so the second table is selected, and
as the result, the font is rejected correctly.

Test: android.graphics.cts.TypefaceTest passed
Change-Id: Ia279aa572094f8b42c893304b8771196c204a945