Verify well-behaved system under heavy disk load.
Abusive or broken apps can go crazy and try allocating all of the
disk space on the device.  This test emulates one of those apps, and
verifies that the user is able to navigate into Settings to uninstall
or free up disk space.

A typical device implementation will use quotactl() to prevent apps
from using an unreasonable amount of disk space.

Also add a test to ensure that the number of inodes is reasonable
compared to the disk size.

Test: cts-tradefed run commandAndExit cts-dev -m CtsAppSecurityHostTestCases -t android.appsecurity.cts.StorageHostTest
Test: cts-tradefed run commandAndExit cts-dev -m CtsOsTestCases -t android.os.cts.EnvironmentTest#testSaneInodes
Bug: 36450358
Change-Id: I74af56d485addf046d58c7d650f5c4d7bbdeeb51