Batterystats CTS test: background wifi scan
Performs wifi scans in both foreground and background processes and
ensures that the counts are incremented correctly.
Note that the previous code for this did not WAI because the
instramentation harness ran everything through a foreground service.

Fixes: 35677312
Test: run cts-dev -m CtsIncidentHostTestCases -t
com.android.server.cts.BatteryStatsValidationTest
Test: Also used st line of
batterystats to ensure that background/foreground usage was as desired.

Change-Id: I2d6d12f84cd3bc236b4db8f5196d0902ab6e9c2d