Test imeOptions attribute in TextView
EditorInfo#imeOptions is an important way for applications to tell how
IMEs should behave on the corresponding text input context.  For
TextView and its sub classes, Android provide a way for application
developers to control values specified to EditorInfo#imeOptions in
layout XMLs.  Here is an example:

  <TextView
      android:id="@+id/my_textview"
      android:layout_width="match_parent"
      android:layout_height="wrap_content"
      android:imeOptions="actionGo|flagNoExtractUi|flagForceAscii" />

This CL introduces basic test cases for the above mechanism to work.

Bug: 28157942
Bug: 37305520
Test: bit CtsWidgetTestCases:android.widget.cts.TextViewTest
Change-Id: I18269ad4ff49fef25513a6992d69741ad75a3d74