Fix shortcut backup CTS
The issue is sometimes the package broadcasts take a long time to be
delivered.  If they arrive after a "bmgr restore" then the shortcut
manager would be confused.  This is exactly what happened after
uninstalling 6 packages in a row.

To avoid this, now we wait until all the broadcast queues drain before
proceeding to the next step.

(Uninstalling all the package at once and wait until the queues drain
didn't work probably because even *posting* the package broadcasts in PMS
is asynchronous using a handler, not just *delivering*?  There's a window
between PMS deciding to send a broadcast, and it showing up in the activity
manager dumpsys.)

Bug: 37423318
Test: cts-tradefed run cts-dev --skip-device-info --skip-preconditions --skip-system-status-check com.android.compatibility.common.tradefed.targetprep.NetworkConnectivityChecker -a armeabi-v7a -l debug -m CtsShortcutHostTestCases -t 'android.content.pm.cts.shortcuthost.ShortcutManagerBackupTest#testBackupAndRestore'
Change-Id: I16fe8a28a361d75088bda825365d511d6a21d7e2