Fix CTS VLoadTest failures under debug context.
Failures were of the form

  Data type mismatch for ElementAt 8 of 4

when running a test.  The problem is that all test
Allocations have signed Elements, but the scripts
expect some to be unsigned.  The fix is to employ
unsigned Elements where necessary to match the scripts.

Test: aosp_x86_64-eng (emulator)
      1) cts -t android.renderscript.cts.VLoadTest with debug.rs.debug 1
         clean
      2) cts -t android.renderscript.cts.VLoadTest with debug.rs.debug 0
         clean

Bug: 37955136
Change-Id: I75b54788aaec4ec41bbade91bcb9d63c1f7eafb7
(cherry picked from commit 994a4fc3f9b8538a6617850459ed0ce9c55e8156)