Verify the default dialer shows HUN on incoming call.
Test: manual
Bug: 35338645

The test is as follows:
1) The user is asked to call the device.
2) If a heads up notification is shown during the incoming call,
and only one notification is shown, the user is asked to check a box.
3) If the box is checked, the user can mark the test as passed.

This test aims to verify that a default dialer, when set,
can detect incoming calls and display a heads up notification.

Additionally, the test verifies that only one notification is shown
for an incoming call, even if more than on dialer application is
installed.

Change-Id: If1dff40c5e185680cdf98248ff5fe62ab41673f9