CTS Keyguard*Tests don't assume screenlock
KeyguardTests and KeyguardTransitionTests work only if
the device has screenlock enabled.

Explicitly enable screenlock as part of the test set up.

See a similar fix in
https://googleplex-android-review.git.corp.google.com/2156477

(This fix is required by the Emulator because it has
screenlock disabled by default.)

Bug: 62088797

Test: This is test code.

This fix was verified on Emulator from emu-master-dev and on
Nexus 5x running bullhead-userdebug 8.0.0 OPR1.170524.002
4041127 dev-keys.

Change-Id: I986b5afb06ad1f42a1c1098d937ea0822c356eef