Fix for CTS nativeopengl tests
EGLCleanupTest makes three calls to eglTerminate() as part of test
setup, to try and ensure that when it later allocates memory between an
eglInitialize() and eglTerminate() that the eglTerminate() actually
frees the memory as it is supposed to. Otherwise the test fails as it
thinks that cleanup is not being performed correctly.

However these calls interfere with the normal hardware accelerated 
rendering code, running on another thread to handle rendering the test 
activity's View, causing its EGL resource to be destroyed, potentially
in the middle of a render, and causing that thread to assert/abort().
The result is that the test can fail even though the cleanup being
tested for is occuring.

This patch simply disables hardware acceleration for the test
application. Note that all the tests are about low level EGL 
functionality, and do not rely on the rendering method.

Bug: 38017204
Test: run cts -p android.nativeopengl
Change-Id: I539cdabf14361b7ad6eee77ece5bd1a62fe18b83