AnimationDrawableTest will fail when AnimationDrawable created from cache
"AnimationDrawableTest#testStart" check the value of "mAnimationDrawable.isOneShot()" on the top of this test and sometimes it finds the value "false".
The return value of "mAnimationDrawable.isOneShot()" is depends on the way how to create the AnimationDrawable from resources or cached one.
When fail occurs, the test sequence will goes like as follows:

1) The CTS framework creat the AnimationDrawable on setBackgroundResource and it might create the object by animationdrawable.xml.
2) The previous test case "testAccessOneShot" will change the value of "mOneShot" to "true".
3) When "AnimationDrawableTest#testStart" is started and the cache of 2) is remained, the object will be initialized from chache, and the value of "mOneShot" is "true" in this case.

Usually, this testcase will pass, but it likely goes fail under system busy.
Please confirm this sequence and fix this.

BUG: 62329931

Change-Id: I9ede3b53ec60eb4f71a989ffc1af58c4925f33b4
Signed-off-by: Takeshi Katae <takeshi.katae.xm@kyocera.jp>