WindowManager: Remove untenable tests.
SurfaceFlinger will throw an error if the WM tries to create
too large of a Surface, but there is no way to predict what this value is.
Practically it should be at least 4096x4096 but there is no guarantee.
It's not enough to query GL_MAX_TEXTURE_SIZE as we don't know what configuration
SurfaceFlinger is using. The CDD has no notes on maximum texture size.

Bug: 62685202
Bug: 63136620
Test: There are fewer now.
Change-Id: I3fd5fe417011a76078f3bc1a6c22ee076d24ea9d