DO NOT MERGE: sphal is another linkernamespace that is available to JNI libs
CTS tries to dlopen every libs under /system/lib and checks if the
result of dlopen is as expected. If a lib name is not mentioned in
/system/etc/public.libraries.txt, then the dlopen should fail with a
message like 'library xxxx is not accessible for the namespace
"classloader-namespace"'. CTS compares the result of dlerror() with this
string literal.

However, there is a corner case that produces different error message;
libX.so is both in /system/lib and /vendor/lib, and libX.so is added to
/vendor/etc/public.libraries.txt. In this case, linker tries to load
/system/lib/libX.so in the 'sphal' namespace (since it uses file's
base name as its soname when crosssing the namespace boundary). As a
result the error message is like '... for the namespace "sphal"'.

So, adding '... for the namsapce "sphal"' as one of the expected error
message for the case.

Bug: 63974074
Test: put libX.so to both /system/lib and /vendor/lib.
add libX.so to /vendor/etc/public.libraries.txt.
android.jni.cts.JniStaticTest#test_linker_namespaces passes.

Merged-In: I89d4d16bc05179fc50eb1b56f97f72b93128d015
Change-Id: I89d4d16bc05179fc50eb1b56f97f72b93128d015