Cleanup platform versions prior to release
We will be running CTS from the 8.0.0 CTS release, so this file
will never be used in a real CTS test, but it is good to clean
it up so it matches the platform version of this release.

Note that because Treehugger checks to make sure that this file
has the PLATFORM_VERSION, I have temporarily added v8.0.0, that
value will go away next week.

Test: Verfied that Treehugger pre-submit passes since that is
the only thing that uses this file.
Bug: 64125438

Merged-In: I22600aee881c4cd89412134f96f8b00a31192f5c
Change-Id: I2a1203b8bd40836ef26b58bd2b1e83e0519fb654