DO NOT MERGE Manually CP ag/1781120 to fix dynamic config file bug
The branch of the original change has diverged greatly from
nougat-mr1-cts-dev, so cosmetic changes were only CPed if doing so
was simple and wouldn't break the build. DO NOT MERGE because of
inevitable merge conflicts with branches containing the original
change, plus inevitable build breakages on internal branches whose
'vendor/xts' project depends on the infra modified in this CL
(The 'vendor/xts' project does not exist for this or other AOSP
branches).

bug: 63030111
Test: run cts-dev -m CtsPreferenceTestCases
Change-Id: Id027f814501f57112c9855711297110f02b1cfdf