Re-enable android.cts.compilation.AdbRootDependentCompilationTest#testCompile_curProfile
Re-enable the test case. It no longer fails after deleting the "hot"
markers from the text profile. This is OK since the format defaults
to hot in the absense of either "startup", "hot", or "post-startup".

The problem with the "hot" markers is that they were not supported
on OC.

Bug: 38462487
Test: cts-tradefed run cts-dev -m CtsCompilationTestCases

(cherry-picked from commit 312924fa1988c9002d63abb6e53dccc756e72559)

Change-Id: I74c76031eb77b35aa79aea422c6c1a493a949abf