Prevent enabling/disabling bluetooth when it is not supported.
The tests will throw exceptions in teardown if the device does not
support bluetooth. If a device has no support for bluetooth a call to
getDefaultAdapter will return null. In those cases we cannot
enable/disable the bluetooth adapter.

We now turn enable/disable to do nothing if no bluetooth adapters is
present. The tests will now pass in the emulator.

Bug: 64713372
Test:  cts-tradefed run cts -m CtsDevicePolicyManagerTestCases -t
       com.android.cts.devicepolicy.DeviceOwnerTest#testBluetoothRestriction

Change-Id: I5fd5e285679d44c7eb9808b13200e1134e0f72d3