Fix JniStaticTest#test_linker_namespaces failure for libz
In a Treble-ized device, there are two libz.so: /system/lib/libz.so and
/system/lib/vndk-sp/libz.so. CTS tries to dlopen both files and the
former should succeed and the latter should fail with an error message
`/system/lib/vndk-sp/libz.so is not accessible for the namespace
"classloader-namespace"`. However, since libz.so is in
/system/etc/public.libraries.txt, which redirects load requests from
classloader-namespace to the (default) namespace, the dlopen fails at
the (default) namespace, but not at the classloader-namespace.

So, the (default) namespace is added to one of the eligible namespaces
for the dlopen error.

Bug: 64950640
Bug: 64888291
Test: android.jni.cts.JniStaticTest#test_linker_namespaces does not
fail on 2017 pixel devices

(cherry picked from commit 855e00e7cfc9355bbc3e959cc24d2ef6b2bb3c29)

Change-Id: I2ef0c7e4d0ce90f2c718636032e5e81a9298ad8b