Do not call removeLockCredentials directly.
This method is an method that relies on the rest of
tearDownLockCredentials to properly clean up the lock state. Calling
it by itself can leave the device still in a potentially locked
state.
This changelist makes removeLockCredentials private and replaces
references to it with tearDownLockCredentials.

Test: run ActivityManagerDisplayTests#testSecondaryDisplayShowWhenLocked
followed by ActivityManagerDockedStackTests#testMinimumDeviceSize.
Ensure both tests path
Fixes: 66944804

Change-Id: Iba61ee5d71fc81a826de592c4f52934f441c0e83