DO NOT MERGE Adjust translucent fullscreen activity component references.
When referring to the activity in the context of the current package,
it is incorrect to fully qualify the activity component. This
changelist updates references to use the shorthand component name
instead, which matches the values reported from the system. This
change also removes the legacy package containing the translucent
activity, which is no longer used.

Change-Id: I69a3f6e94f035d5a5bff39428d46b0cf656e0a08
Fixes: 68826398
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerAppConfigurationTests#testNonfullscreenAppOrientationRequests
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerAppConfigurationTests#testNonFullscreenActivityPermitted