Fix too strict requirements in IME state handling
Currently the following two tests essentially require the system to
always pick up the previously selected system IME when the CTS started
after disabling or uninstalling the current IME, which is actually
not guaranteed.

 - InputMethodServiceLifecycleTest#testDisableCurrentIme
 - InputMethodServiceLifecycleTest#testUninstallCurrentIme

It's true that the system tries to pick up an appropriate system IME
when the current one becomes unavailable, but which system IME will
be picked up is not deterministic.

This CL makes it clear and only requires that the new IME is not the
uninstalled/disabled one.

Fixes: 69276863
Test: atest CtsInputMethodServiceHostTestCases
Change-Id: I2b6563e9276e380dca485dbe5e3782f09e0dc445