Try to deflake CtsInputMethodServiceHostTestCases
Although this CL makes CtsInputMethodServiceHostTestCases a bit slow
again (15s -> 22s on taimen-userdebug), major flakiness is expected to
be addressed with this CL.

Imagine the following steps:
  1. Install an APK A that contains IME X.
  2. Try to enable the IME X.
  3. Try to select the IME X as the current IME.

The problem is that there is always a delay between when the APK A is
installed and when the IME X is recognized by the
InputMethodManagerService (IMMS).

To makes the situation worse, our previous test implementation relied
on "adb shell ime enable <IME ID>" command to enable the IME X, which
does not work until IMMS actually recognizes the IME X.

With this CL, the above step 2 is moved from host-side to device-side
so that we can check what IMEs are recognized by IMMS without relying
on "adb shell ime" commands, since for application developers, the
following two methods are the source of truth.
 * InputMethodManager#getInputMethodList()
 * InputMethodManager#getEnabledInputMethodList()

This CL also does the step 2 with Settings.Secure#putString() instead
of relying on "adb shell ime enable" command, because it is the real
world use case for users and developers.  Note that CTS test APKs is
actually able to have "WRITE_SECURE_SETTINGS" because it is defined to
be "development" protection-level.

Bug: 70730589
Test: atest CtsInputMethodServiceHostTestCases
Change-Id: I306d66a65ebabefeec02dd247f26f87709731a27