Add dependency on android.test.mock.stubs for Proguard
Even though the affected target does not require the library to compile
it does need it for Proguard.

The following change descriptions were generated automatically and so
may be a little repetitive. They are provided to give the reviewer
enough information to check the comments match what has actually been
changed and check the reasoning behind the changes.

* apps/CtsVerifier/Android.mk
    Added 'android.test.mock.stubs' to LOCAL_JAVA_LIBRARIES because
    without it Proguard fails while processing CtsVerifier because it
    needs, but cannot find classes from android.test.mock.stubs.

Bug: 30188076
Test: make dist
Change-Id: I6e6dc37fc3a5f188a9a9991374864b221ceff7f4