Allow MockIme to set/unset window flags
This is a preparation CL to fix Bug 69002467 with tests.

This is a follow up CL to my previous CL [1], which introduced an
ability to specify any window flag to Mock IME during tests.  One
major limitation in my previous CL is that there was no way to remove
a set of window flags.

With this CL, now the client can specify the mask bits to teach Mock
IME about which flags need to be cleared.

  // Make sure that FLAG_LAYOUT_IN_OVERSCAN flag is unset.
  try(MockImeSession imeSession = MockImeSession.create(
        context, uiautomation,
        new ImeSettings.Builder().setWindowFlags(
                0, FLAG_LAYOUT_IN_OVERSCAN)) {

      // Use imeSession to interact with the Mock IME here.

  }

 [1]: I6bcab949dddc50a9a5cbd0a59b205581d344e46d
      03e0f654c09a4a033420e8b7698ed8833855816f

Bug: 69002467
Test: atest CtsInputMethodTestCases
Change-Id: I76b6e7aed37235ca3fba0072e9e524260a210fd5